backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
pymupdf
google-generativeai